Map reverse lookup: Use English to map back to DU country topics
We are looking up DU topics for countries by name,
contructing the ID from "geo" + country name. Of course,
that only works for English.
HACK quick fix is to use Nominatim in English.

Better fix is to record country ISO codes in DU as properties,
and look these up. This needs more infrastructure for arbitrary
properties in UniNav, though.